pyproject.toml
[project]
name = "djchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.1",
    "cryptography>=46.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.2.1",
]
